+++ src/core/Jalasoft.TeamUp.Projects.Core/InvitationsService.cs
@@ -7,14 +7,18 @@
     using Jalasoft.TeamUp.Projects.Core.Validators;
     using Jalasoft.TeamUp.Projects.DAL.Interfaces;
     using Jalasoft.TeamUp.Projects.Models;
+    using Jalasoft.TeamUp.Projects.Models.Enums;
 
     public class InvitationsService : IInvitationsService
     {
         private readonly IInvitationsRepository invitationsRepository;
+        private readonly IProjectsRepository projectsRepository;
+        private IInvitationsRepository @object;
 
-        public InvitationsService(IInvitationsRepository invitationsRepository)
+        public InvitationsService(IInvitationsRepository invitationsRepository, IProjectsRepository projectsRepository)
         {
             this.invitationsRepository = invitationsRepository;
+            this.projectsRepository = projectsRepository;
         }
 
         public Invitation[] GetInvitationsByProjectId(string projectId)
@@ -37,6 +41,20 @@
         {
             InvitationValidator validator = new InvitationValidator();
             validator.ValidateAndThrow(invitation);
+            var invitationStatus = InvitationStatus.Accepted;
+            if (invitation.Status == invitationStatus.ToString())
+            {
+                Contact contact = new Contact
+                {
+                    IdResume = invitation.ResumeId,
+                    Name = invitation.ResumeName
+                };
+
+                var project = this.projectsRepository.GetById(Guid.Parse(invitation.ProjectId));
+                project.MemberList.Add(contact);
+                this.projectsRepository.UpdateById(project);
+            }
+
             var result = this.invitationsRepository.UpdateById(invitation);
             return result;
         }
