--- src/core/Jalasoft.TeamUp.Projects.Core/Validators/InvitationsStatus.cs
@@ -1,9 +1,0 @@
-﻿namespace Jalasoft.TeamUp.Projects.Core.Validators
-{
-    public enum Status
-    {
-        Invited = 0,
-        Accepted = 1,
-        Rejected = 2,
-    }
-}
