+++ src/dal/Jalasoft.TeamUp.Projects.DAL/InvitationsMongoDbRepository.cs
@@ -13,7 +13,6 @@
         private static MongoClient client;
         private static IMongoDatabase database;
         private static IMongoCollection<Invitation> collection;
-        private readonly IProjectsRepository projectsRepository;
 
         public InvitationsMongoDbRepository()
         {
@@ -58,15 +57,6 @@
         public Invitation UpdateById(Invitation invitation)
         {
             var filter = Builders<Invitation>.Filter.Eq(s => s.Id, invitation.Id);
-            if (invitation.Status == "Accepted")
-            {
-                Contact contact = new Contact();
-                contact.IdResume = invitation.ResumeId;
-                contact.Name = invitation.ResumeName;
-                var project = this.projectsRepository.GetById(Guid.Parse(invitation.ProjectId));
-                project.MemberList.Add(contact);
-                this.projectsRepository.UpdateById(project);
-            }
 
             var proj = collection.ReplaceOne(filter, invitation);
