+++ src/models/Jalasoft.TeamUp.Projects.Models/Enums/InvitationStatus.cs
@@ -1,0 +1,9 @@
+﻿namespace Jalasoft.TeamUp.Projects.Models.Enums
+{
+    public enum InvitationStatus
+    {
+        Invited = 0,
+        Accepted = 1,
+        Rejected = 2,
+    }
+}
